# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation jardin de cocagne",
        "version": "0.1"
    },
    "paths": {
        "/api/adherents/update": {
            "post": {
                "tags": [
                    "Adherents"
                ],
                "summary": "Update user's profile",
                "description": "Updates the user's profile information",
                "operationId": "updateProfile",
                "requestBody": {
                    "description": "User's profile data for update",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "raison_sociale": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "civilite": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "enum": [
                                            "mme",
                                            "mr"
                                        ]
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "ville": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "code_postal": {
                                        "type": "numeric"
                                    },
                                    "numero_telephone": {
                                        "type": "numeric",
                                        "pattern": "/^0[1-9]\\d{8}$/"
                                    },
                                    "numero_telephone2": {
                                        "type": "numeric",
                                        "pattern": "/^0[1-9]\\d{8}$/",
                                        "nullable": true
                                    },
                                    "numero_telephone3": {
                                        "type": "numeric",
                                        "pattern": "/^0[1-9]\\d{8}$/",
                                        "nullable": true
                                    },
                                    "profession": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "maxLength": 2048,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Error updating profile"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/adherents/store": {
            "put": {
                "tags": [
                    "Adherents"
                ],
                "summary": "Store a new adherent",
                "description": "Stores a new adherent with the provided data",
                "operationId": "storeAdherent",
                "requestBody": {
                    "description": "Adherent data for storage",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "raison_sociale": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "civilite": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "ville": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "code_postal": {
                                        "type": "numeric"
                                    },
                                    "numero_telephone": {
                                        "type": "numeric",
                                        "pattern": "/^0[1-9]\\d{8}$/"
                                    },
                                    "numero_telephone2": {
                                        "type": "numeric",
                                        "pattern": "/^0[1-9]\\d{8}$/",
                                        "nullable": true
                                    },
                                    "numero_telephone3": {
                                        "type": "numeric",
                                        "pattern": "/^0[1-9]\\d{8}$/",
                                        "nullable": true
                                    },
                                    "profession": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "nullable": true
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "maxLength": 2048,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Adherent stored successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Error storing adherent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/adherents/{id}": {
            "get": {
                "tags": [
                    "Adherents"
                ],
                "summary": "Get details of a specific adherent",
                "description": "Returns details of a specific adherent by ID",
                "operationId": "getAdherentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the adherent to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Adherent not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Adherents"
                ],
                "summary": "Delete a specific adherent",
                "description": "Deletes a specific adherent by ID",
                "operationId": "deleteAdherent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the adherent to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Adherent deleted successfully"
                    },
                    "404": {
                        "description": "Adherent not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}